Rework simulate on cycle invoice

+++ packages/server/customer-os-api/graph/resolver/invoice.resolvers_it_test.go
@@ -101,7 +101,7 @@ func TestInvoiceResolver_Invoice(t *testing.T) {
 	require.Equal(t, 1, len(invoice.InvoiceLineItems))
 	require.Equal(t, "SLI 1", invoice.InvoiceLineItems[0].Description)
 	require.Equal(t, 100.0, invoice.InvoiceLineItems[0].Price)
-	require.Equal(t, 1, invoice.InvoiceLineItems[0].Quantity)
+	require.Equal(t, int64(1), invoice.InvoiceLineItems[0].Quantity)
 	require.Equal(t, 100.0, invoice.InvoiceLineItems[0].Subtotal)
 	require.Equal(t, 19.0, invoice.InvoiceLineItems[0].TaxDue)
 	require.Equal(t, 119.0, invoice.InvoiceLineItems[0].Total)
